refactor: move syntax analysis code into src/syntax-analysis
