+++ src/syntax-analysis/syntax-type.ts
