+++ src/syntax-analysis/token.ts
@@ -1,0 +1,8 @@
+import Syntax from "./syntax-type";
+
+
+export default interface Token {
+    syntax: Syntax;
+    lexeme: string;
+    value: any;
+}
